+++ src/tr2/game/viewport.c
@@ -120,14 +120,7 @@ static void M_ApplyGameVars(const VIEWPORT *const vp)
 
 void Viewport_Reset(void)
 {
-    int32_t win_width;
-    int32_t win_height;
-    if (Shell_IsFullscreen()) {
-        win_width = Shell_GetCurrentDisplayWidth();
-        win_height = Shell_GetCurrentDisplayHeight();
-    } else {
-        SDL_GetWindowSize(g_SDLWindow, &win_width, &win_height);
-    }
+    const SHELL_SIZE size = Shell_GetCurrentSize();
 
     VIEWPORT *const vp = &m_Viewport;
     switch (g_Config.rendering.aspect_mode) {
@@ -138,12 +131,12 @@ void Viewport_Reset(void)
         vp->render_ar = 16.0 / 9.0;
         break;
     case AM_ANY:
-        vp->render_ar = win_width / (double)win_height;
+        vp->render_ar = size.w / (double)size.h;
         break;
     }
 
-    vp->width = win_width / g_Config.rendering.scaler;
-    vp->height = win_height / g_Config.rendering.scaler;
+    vp->width = size.w / g_Config.rendering.scaler;
+    vp->height = size.h / g_Config.rendering.scaler;
     if (g_Config.rendering.aspect_mode != AM_ANY) {
         vp->width = vp->height * vp->render_ar;
     }
@@ -171,9 +164,8 @@ void Viewport_Reset(void)
     M_InitGameVars(&m_Viewport);
     M_ApplyGameVars(&m_Viewport);
 
-    const int32_t win_border = win_height * (1.0 - g_Config.rendering.sizer);
-    Render_SetupDisplay(
-        win_border, win_width, win_height, vp->width, vp->height);
+    const int32_t win_border = size.h * (1.0 - g_Config.rendering.sizer);
+    Render_SetupDisplay(win_border, size.w, size.h, vp->width, vp->height);
 }
 
 const VIEWPORT *Viewport_Get(void)